pos-print: move core classes from client to core package
The classes which are used by core (Receipt/ReceiptItem) are now moved
from pos-print-client to pos-print's core package as the client will
have to use DTO object instead as they should not be coupled.

Few minor fixes are also performed by this change:
 * removed some un-used classes
 * CommandCLI is now fixed and works as expected
 * kotlin version is updated to 1.1.1
 * CoreModule is now used and bindings are moved there from the HTTP one
