+++ pos-print-acceptance-tests/build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-  ext.kotlin_version = '1.1.0'
+  ext.kotlin_version = '1.1.1'
   repositories {
     mavenCentral()
   }
