+++ pos-print/src/main/java/com/clouway/pos/print/adapter/http/HttpModule.java
@@ -26,7 +26,5 @@ public class HttpModule extends AbstractModule {
     };
 
     install(pageBricks);
-
-    bind(PrinterFactory.class).to(FP705PrinterFactory.class);
   }
 }
